+++ lib/flutter_webview_plugin_ios_android.dart
@@ -1,4 +1,4 @@
-library flutter_webview_plugin;
+library flutter_webview_plugin_ios_android;
 
 export 'src/base.dart';
 export 'src/javascript_channel.dart';